refactor : 폴더 depth 제거
