+++ src/app/(main)/profile/setting/delete/components/DeleteSection.tsx
@@ -1,7 +1,7 @@
 'use client';
 import DeleteHeader from './DeleteHeader';
-import Step1 from './step1/Step1';
-import Step2 from './step2/Step2';
+import Step1 from './Step1';
+import Step2 from './Step2';
 import { useFunnel } from '@/hooks/useFunnel';
 import { useRouter } from 'next/navigation';
